New fetcher class for easier getting objects from db and basic deposit/withdrawal functionality

+++ personalaccount.h
@@ -1,13 +1,12 @@
 #ifndef PERSONALACCOUNT_H
 #define PERSONALACCOUNT_H
-#include "account.h"
 
 class PersonalAccount
 {
 public:
     int id;
     int user_id;
-    double balance;
+    int balance;
 };
 
 #endif // PERSONALACCOUNT_H
